requirements.txt
alembic==1.7.7
APScheduler==3.9.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
configparser==5.2.0
cryptography==37.0.2
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mysqlclient==2.1.0
packaging==21.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
